rearrange handler tests
